requirements.txt
Flask==3.1.3
flask-cors==6.0.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
werkzeug==3.1.7
pymongo==4.6.1
python-dotenv==1.0.0
flask-mail==0.9.1
PyJWT==2.8.0
bcrypt==4.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
